+++ VERSION_HISTORY.md
@@ -138,7 +138,7 @@
     * Gate enable/disable routines
     * Gate enable/disable example
 * Counter2 mode0 routines implemented
-    * TODO counter2 mode0 examples    
+    * counter2 mode0 examples    
 * TODO Header based implementation for counter0 mode1
 * TODO Header based implementation for counter0 mode2
 * TODO Header based implementation for counter0 mode3